Improve "end turn" action handling.
The recently added KStandardGameAction::endTurn() action now has an own
default shortcut of Ctrl+E. The button is added to the toolbar by default.

Furthermore, the "end turn" button now is enabled/disabled depending on the
game state. It no longer is possible to end the turn while in the middle of
a send fleet command sequence.

Use the very same condition for enabling/disabling the custom "end turn"
button and the "end turn" action. This required some code cleanup.

The custom palette for the game board now colours disabled elements darker
so that they actually look disabled.

REVIEW: 113528
GUI: