Fix "end turn" action.
The "end turn" button now is enabled/disabled depending on the game state.
It no longer is possible to end the turn while in the middle of a send
fleet command sequence.

Use the very same condition for enabling/disabling the custom "end turn"
button and the "end turn" action.

The custom palette for the game board now colours disabled elements darker
so that they actually look disabled.

REVIEW: 113598