Make "current standings" dialogue dockable.
The dialogue "current standings" is replaced by a dockable widget. This way
the player is able to see the standings all the time without the need to
open/close the dialogue at the beginning of each turn.

The dockable widget can also be made floating, so it can be used almost the
same as the modal dialogue before.

REVIEW: 113782
GUI: