Disable planet owner combo box if required.
Changing the owner of a planet in new game dialogue currently crashes if
the selected planet moves off the map due to a resize. Steps to reproduce
are in the bug report. This patch makes owner combo box follow same rules
for enabling as production and kill percentage.

Patched by Michael Reeves.
This is implemented in Konquest 2.1.

BUG: 328236
FIXED-IN: 4.11
REVIEW: 114232