Properly check for a valid selection in minimap.
Changing the dimensions of the map in game setup dialogue recreates the
game board. The sector selection is kept. However, the selection might
become invalid if the game board shrinks and the selection is not on the
board anymore. Adding a new player to the game will then result in
checking a sector outside the game board (NULL is returned), so that
a method is invoked on a NULL pointer, leading to a crash - see
NewGameDlg::slotNewOwner().

This is implemented in Konquest 2.1.

BUG: 329163
FIXED-IN: 4.11.5
(cherry picked from commit 238c8f3f12c84107be9095e63b135e8c94786f00)