Update fleet overview dialogue.
The fleet overview dialogue is not resized to a constant size anymore, and
automatically adopts to the number of (own) attack fleets. Numbers shown in
that dialogue are handled as numbers now rather than as strings so that
they are sorted properly (1, 2, 10, 20 instead of 1, 10, 2, 20). They are
as well aligned to the right so that they are nicer to read. The source and
destination planet fields show the player colour of the owner now.

The structure resembles the one of the standings widget and this indeed
is a first step of converting the dialogue into a dockable widget.

REVIEW: 114735
GUI: