Center board vertically if required.
The board is centered horizontally already in case the board is tall. If
it is wide however, it still is aligned to the top leaving empty space
just at the bottom. This patch vertically centers a wide board.

Remove a work-around for Qt 4.6; Qt 4.8 is required and there is no
visible difference without that work-around.

The element "screen" had a transparent bar at the upper edge and
adjustments were done in the code to hide it. As the element has been
fixed, the adjustments can be removed.

REVIEW: 114851
GUI: