Update default theme SVGZ.
The element "screen" has a transparent bar at the upper edge. The code
tries to hide this bar by adjusting the shown area.

This patch removes that transparent bar from the element. As the board is
always aligned to the top of the viewport, a change is not visible right
now. However, the SVGZ file now is compressed, so its size now is about
80 kB rather than 1274 kB.

Thanks to Ian Wadham for the updated theme.